cruddddd

+++ unihive/forum/views.py
@@ -1,18 +1,18 @@
-from django.shortcuts import render, redirect, get_list_or_404
+from django.shortcuts import get_object_or_404, render, redirect, get_list_or_404
 from django.urls import reverse
 from .models import Task
 from .forms import TaskForm
 
 # Create a task
-def task_create(request):
+def create_task(request):
     if request.method == "POST":
         form = TaskForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect(reverse("forum:task_list"))
+            return redirect(reverse("create_task"))
     else:
         form = TaskForm()
-    return render(request, "forum/task_create.html", {"form": form})
+    return render(request, "forum/create_task.html", {"form": form})
 
 
 # Retrieve task list
@@ -21,27 +21,25 @@ def task_list(request):
     return render(request, "forum/task_list.html", {"tasks": tasks})
 
 
-# Retrieve task detail
-def task_detail(request, pk):
-    task = Task.objects.get(pk=pk)
-    return render(request, "forum/task_detail.html", {"task": task})
-
-
 # Update a task
-def task_update(request, pk):
-    task = get_list_or_404(pk=pk)
+def update_task(request, pk):
+    task = get_object_or_404(Task, pk=pk)
     if request.method == "POST":
         form = TaskForm(request.POST, instance=task)
         if form.is_valid():
             form.save()
-            return redirect(reverse("forum:task_list"))
+            return redirect("task_list")
     else:
         form = TaskForm(instance=task)
-    return render(request, "forum/task_update.html", {"form": form})
+    return render(request, "forum/update_task.html", {"form": form})
 
 
 # Delete a task
-def task_delete(request, pk):
-    task = get_list_or_404(pk=pk)
-    task.delete()
-    return redirect(reverse("forum:task_list"))
+def delete_task(request, pk):
+    task = get_object_or_404(pk=pk)
+    if request.method == 'POST':
+        task.delete()
+        return redirect("task_list")
+    return render(request, 'forum/delete_task.html', {'task': task})
+
+
